backend/requirements.txt
flask
flask-sqlalchemy
flask-cors==3.0.10
psycopg2-binary
numpy
pandas
pillow
matplotlib